feat: extract parsing utils to a separate module
